Adds local tracking of opened previews
Previously, there was not way to know if an album had been previewed.

This commit adds the ability to track opened previews. This tracking is done with `window.localStorage`. The history is tracked with a clickable element to the right of the preview button. This clickable element will change state (previewed, represented by a color change) when a preview is initiated. Additionally, the preview state can be overridden by directly clicking this element.