Modifies preview window width
The preview window width was set to 100%. This resulted in difficulty using the playbar control in the iFrame. I looks like modifying the CSS in the iFrame could allow for a custom width for the playbar BUT modifying elements in the iFrame is not easy. Thus, in this commit the preview window is simply expanded to 150%.