track page loads for albums (#20)
* move all functions out of 'main' for cleanup.

* add preview state cache for when previewing album on album page.

Previously, the data storage mechanism used did not allow for cross-domain caching of previews. Specifically, this meant that it was not possible to store an album page load and then update the preview marker on a label page.

By moving to the `chrome.storage` API, this cross-domain storage is possible. So now, when opening an album's page, that history will be stored and used on the label page later.

This API is used in `sync` mode and will allow for history to be work across multiple machines.

Migration from the old storage backend is also automatically handled.

additionally:
* moved all functions out of "main"
* change page loaded check to jquery document ready check
* set jquery to specific version (3.3.1)