add volume control to album pages
There is a volume control for HTML5 audio elements but Bandcamp just doesn't use it. This feature adds a control to the right side of the player on album pages to provide volume control.

The control uses a css file to make the `input` element (which controls the volume) to look as close to the player playbar as possible.

Note: one element of the playbar which is currently not replciated is that the "playhead" graphic is slightly larger than the timeline graphic. This has been sacrificed with the `input` element in order to replicate the white/grey effect of aboe/below the "playhead" (which is done as a hack with the `boxshadow` + `overflow: visible`)