update implementation of db accessor to allow for testing without real db instance
Previously tests were not very well isolated because they used a running instance of IDB. Ideally, `Sinon` would be able to stub the imported module for the DB accessor BUT this does not appear to work with WebPack.

To provide proper test fixturing, a new class `DBUtils`, was created which acts as a thin wrapper around IDB and provides a method for returning a db handle. This was previously done with a standalone function but was wrapped in a class to allow an IDB stub to be used in tests.

This new `DBUtils` class is now injected into the `LabelViewBackend` static methods for accessing the IDB instance. And, because of this injection, the tests for `LabelViewBackend` are now completely isolated from any real DB instance.

additionally:
* run linter