add ConfigBackend class to handle receiving configs, syncronizing configs to DB, and broadcasting synced config
This class is initially used to extend the Waveform class (also in this commit) to allow the Waveform toggle selection to be stored and persist between different pages and on reload.

This class was designed more generally to store config data and this will also allow for a "configuration" page in the future where the toggle--and other configurations such as turning off specific features--can be edited.