Feat/global config (#79)
* update implementation of db accessor to allow for testing without real db instance

Previously tests were not very well isolated because they used a running instance of IDB. Ideally, `Sinon` would be able to stub the imported module for the DB accessor BUT this does not appear to work with WebPack.

To provide proper test fixturing, a new class `DBUtils`, was created which acts as a thin wrapper around IDB and provides a method for returning a db handle. This was previously done with a standalone function but was wrapped in a class to allow an IDB stub to be used in tests.

This new `DBUtils` class is now injected into the `LabelViewBackend` static methods for accessing the IDB instance. And, because of this injection, the tests for `LabelViewBackend` are now completely isolated from any real DB instance.

* add ConfigBackend class to handle receiving configs, syncronizing configs to DB, and broadcasting synced config

This class is initially used to extend the Waveform class (also in this commit) to allow the Waveform toggle selection to be stored and persist between different pages and on reload.

This class was designed more generally to store config data and this will also allow for a "configuration" page in the future where the toggle--and other configurations such as turning off specific features--can be edited.

additionally:
* run linter