fix bug causing Waveform to not work when Checkout is running
Each of the classes was creating its own Chrome port for communication. It appears that only one of those can receive messages from the `config_backend` script. To fix this, a single port is created in `main` and passed to each object that needs access to config.

This pattern makes sense and should be how all classes get access configuration data in the future.