Feature/145 import export cart (#195)
add ability to import and export cart

this is done by providing buttons attached to the cart item on any page. The export will generate a JSON file and the import will handle this JSON file.

additionally -- the reset button from the 1-click buy feature has been moved over to this component. If the cart has been modified -- exporting is not allow and the reset button is show. Otherwise, the reset button is hidden and the export button is visible

additionally:
* move download file function to utilits
* move dateString function to utilities
* update date format for cart and downloads to be yy-mm-dd
* move re-usable static method to component and re-name component to be for button types
* update download_helper and tests for previous changes
* fix potential failure mode of Player and add test
* update Player 1-click feature with a small fix and add test